requirements.txt
streamlit==1.38.0
pandas==2.2.2
mysql-connector-python==9.0.0
folium==0.17.0
streamlit-folium==0.20.0
reportlab==4.2.2
Pillow==10.4.0
requests==2.32.3
PyDrive2==1.20.0
protobuf==5.28.2
python-dateutil==2.9.0.post0
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
oauth2client